library/templates/adk-agent/requirements.txt
google-adk
python-dotenv
rich
prompt_toolkit
mcp
pyyaml